Fixed rounding bug that caused raising a users volume to not change it, and lowering it to irrevocably set it to 0

+++ ProxChatClientGUI/ProxChat.cs
@@ -505,7 +505,7 @@ namespace ProxChatClientGUI
 
         private void OnChangeVolume(string username, byte newVolume)
         {
-            float percentage = newVolume / Settings.Instance.VolumePrefs![username];
+            float percentage = ((float)newVolume) / Settings.Instance.VolumePrefs![username];
             model.AddMessage(() =>
             {
                 model.RecalculateRealVolume(username, percentage);
